Tests/LibWeb/Text/expected/clipboard-text-block-boundary-newlines.txt

adjacent blocks: "one\ntwo\nthree"
single block: "only"
br: "a\nb"
